Split up user lib tests
Resf #340
