+++ tests/lib/user/test_user.py
@@ -4,23 +4,19 @@
 # the Apache License, Version 2.0. See the LICENSE file for details.
 from datetime import timedelta, timezone, datetime
 
-import pytest
-
 from pycroft import config
 from pycroft.helpers.i18n import localized
 from pycroft.helpers.interval import closedopen, single
 from pycroft.lib import user as UserHelper
 from pycroft.lib.facilities import get_room
-from pycroft.lib.user import move, move_out, move_in, edit_address
+from pycroft.lib.user import move, move_out, move_in
 from pycroft.model import (
     session, host)
-from pycroft.model.user import User, RoomHistoryEntry
 from tests import FactoryWithConfigDataTestBase, FactoryDataTestBase
 from tests.factories import UserWithHostFactory, MembershipFactory, UserFactory, \
     RoomFactory, ConfigFactory
 from tests.factories.address import AddressFactory
 from ... import factories
-from ...factories.user import UserWithoutRoomFactory
 
 
 class Test_User_Move(FactoryDataTestBase):
@@ -438,88 +434,3 @@ class UserRoomHistoryTestCase(FactoryDataTestBase):
         self.assertIsNone(rhe.ends_at)
 
 
-class TestUserChangeAddressTestCase(FactoryDataTestBase):
-
-    def create_factories(self):
-        super().create_factories()
-        self.admin = UserFactory()
-
-    @property
-    def address_args(self) -> dict[str]:
-        return {
-            'street': "Blahstraße",
-            'number': "5",
-            'addition': "Keller",
-            'zip_code': "01217",
-            'city': "Dresden",
-            'state': None,
-            'country': None,
-        }
-
-    # TODO with pytest, make this a parametrized test
-    def assert_user_address_change(self, user: User, address_args: dict[str]):
-        assert not user.has_custom_address
-
-        edit_address(user, self.admin, **address_args)
-        self.session.commit()
-
-        self.session.refresh(user)
-        if user.room:
-            assert user.has_custom_address
-        for key, val in self.address_args.items():
-            if key == 'country':
-                assert user.address.country == val or 'Germany'
-                continue
-            assert getattr(user.address, key) == (val or '')
-        assert len(user.log_entries) > 0
-        log_entry = user.log_entries[-1]
-        assert log_entry.author == self.admin
-
-    def test_plain_user_address_add(self):
-        self.assert_user_address_change(UserWithoutRoomFactory(), self.address_args)
-
-    def test_user_with_room_different_address(self):
-        self.assert_user_address_change(UserFactory(), self.address_args)
-
-
-## NAME SIMILARITY TESTS
-THRESHOLD = 0.6
-
-@pytest.mark.parametrize("one,other", [
-    ("Hans", "Hans_"),
-    ("  Franz", "Franz"),
-    ("Tobias Fuenke", "Tobias Fünke"),
-    ("Tobias Fünke", "Tobias"),
-    ("Richard Paul Astley", "Rick Astley"),
-])
-def test_similar_names(one, other):
-    assert UserHelper.are_names_similar(one, other, THRESHOLD)
-
-
-@pytest.mark.parametrize("one,other", [
-    ("Hans", "Definitiv ganz und gar nicht Hans"),
-    ("Tobias Fuenke", "Fünke Tobias"),
-    ("Lukas Juhrich der Große", "Lucas der geile Hecht"),
-])
-def test_nonsimilar_names(one, other):
-    assert not UserHelper.are_names_similar(one, other, THRESHOLD)
-
-
-class SimilarUserTestCase(FactoryDataTestBase):
-    def create_factories(self):
-        super().create_factories()
-        # We need a user in the same room
-        self.room = factories.RoomFactory()
-        self.similar_user_this_room = factories.UserFactory(room=self.room, name="Tobias Fuenke")
-
-        self.similar_user_room_history = factories.UserFactory(name="Tobias")
-        self.session.add(RoomHistoryEntry(room=self.room, user=self.similar_user_room_history))
-
-        # nonsimilar users (same room / room history)
-        factories.UserFactory(room=self.room, name="Other dude")
-        self.session.add(RoomHistoryEntry(room=self.room,
-                                          user=factories.UserFactory(name="Other dude")))
-
-    def test_similar_users_found(self):
-        assert UserHelper.find_similar_users("Tobias Fünke", self.room, THRESHOLD) \
-               == [self.similar_user_this_room, self.similar_user_room_history]
